--- src/components/Input.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react'
-
-class Input extends Component {
-  render () {
-    return (
-      <input type='' name='' />
-    )
-  }
-}
-
-export default Input
